// src/repositories/common/types.ts
export interface IPerson {
  id: string;
  firstName: string;
  middleName: string;
  lastName: string;
  title: string;
  email: string;
  dodid: string;
  rank: string;
  branch: string;
  phone: string;
  address: string;
  dutyPhone: string;
  dutyTitle: string;
  // primaryOrganizationId: 3fa85f64-5717-4562-b3fc-2c963f66afa6,
  // organizationMemberships: [
  //   3fa85f64-5717-4562-b3fc-2c963f66afa6
  // ],
  // organizationLeaderships: [
  //   3fa85f64-5717-4562-b3fc-2c963f66afa6
  // ]
}